testipm/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Cython==0.29.14
Deprecated==1.2.10
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
gensim==3.8.3
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.10
ipaddress==1.0.23
jmespath==0.10.0
joblib==0.16.0
Js2Py==0.70
mail-parser==3.12.0
nltk==3.5
numpy==1.19.1
packaging==20.4
pipwin==0.5.0
playsound==1.2.2
psycopg2==2.8.5
PyAudio
pydub==0.24.1
PyGithub==1.51
pyjsparser==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
simplejson==3.17.0
six==1.14.0
smart-open==2.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
tqdm==4.48.0
tzlocal==2.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
